requirements.txt
torch
torchvision
numpy
pandas
matplotlib
tqdm
ipykernel
pyyaml
